Rework udev rule some more
Now that the rules to remove device properties when the CD is ejected
have been fixed, another bug in these rules shows up: there are 2 rules
very similar to the rules for disk insertion. One of them matches
ID_FS_TYPE != "iso9660" and the other matches ID_FS_TYPE != "udf" which
means that at least one of them will trigger and will remove the
OSINFO_* properties that have been just set by the insertion rules.

While attempting to fix them, I've noticed that these properties go
away anyway when the disk is removed and this has been confirmed on
 #udev, so we can get rid of them.  While at it, I've added a test to
check that the 'change' events correspond to a media change.