install script: generates the output in a file
Add variant of osinfo_install_script_generate() that outputs the
script into a file in the given directory. Also add API to specify a
custom prefix for generated file.

Moreover, the "filename" attribute was added in the "template" element
in the install script data, and it will be used as the output filename
for each script and will be more detailed below. Also, support to this
field was provided in the XML schema and in the osinfo_loader as well.

About the osinfo-install-script tool:
    If the prefix argument is NULL, the output files will be written as
    set in data/install-scripts/*.xml (in filename attribute from
    template element):
        - Linuxes: fedora.ks
        - Windows 2k3r2 and older: windows.sif
        - Windows 2k8 and newer: windows.xml
    Otherwise, the prefix will be prepended in the filename as:
    <prefix>-<filename>

    If the dirname argument is NULL, the output files will be written in
    the current directory.

It will be used to create, easily, multiple scripts, as used in:
http://bugzilla-attachments.gnome.org/attachment.cgi?id=214681

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>