Add desktop profile for Windows
To create an user and set their avatar, we need to create 2 new files,
that will be used in older Windows.
(http://bugzilla-attachments.gnome.org/attachment.cgi?id=214681)

Was discussed in the ML about don't create this files and put all
information in only one script, separating each file in one template, as
suggested in:
https://www.redhat.com/archives/virt-tools-list/2012-June/msg00078.html

In the implementation time, a limitation occurred, and I realized that
I was doing the things in the wrong way. (if there is 1 script for all
the files and _generate*() gives you 1 file as output, how we could
generate different files from the unique file?) So, I would like to go
back with the first implementation, creating different files for each
file used by windows xp (e. g.)

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>