fedora,installer: Don't drop 'base' package group
In commit e625641, we dropped 'base' package group from destkop profile
script thinking that this essential package is pulled as dep by other
groups. This turned out to be not true and Fedora 18 documentation states
that this package must be explicitly requested.

So instead, we now do the same as we do in jeos profile: Use 'standard'
group if dealing with >= F18, 'base' otherwise.