API to query signed status of device drivers
Some OS vendors recommend or require device drivers to be signed by them
before these device drivers could be installed on their OS. An API to
query signed status of the device driver will be useful for apps to be
able to make a decision whether they want to use the driver or not.

Later patches add API for querying signature requirements from a script
and to possibly disable these checks.