win8.1: Disable unattended installation
When win8.1 entry was added it was mostly a c&p from a previous
version and the installer part was mistakenly also copied. Since
win8.1 unattended installation has never been tested and some people
have reported it not to work, lets disable it for now and re-enable
when we assure it works.