debian: Add needs-internet flag to JEOS script
Maybe it is possible to avoid the need of an internet connection in this
script but in the current version debian will notify the user with a
"bad archive mirror" message if the connection is not present. This
question seems not to be meant to be preseeded and I don't know a way
to do it.