debian: Better resources for wheezy
We currently specify recommended RAM for wheezy to be 512MiB and storage
to be 5.4GB. Given that the user might not only want to install a bare
system but also do something with the system 20GB of disk space and 1GiB
of RAM are a reasonable amount. Furthermore the debian desktop express
installation script will fail with less than about 15GB or so.