oses: Add Fedora21
This just adds a plain Fedora21 entry. F21 is going to have different
versions like 'Workstation' and 'Server' but we don't seem to distinguish
for other distros so I just stuck with the pattern here.

This currently doesn't work for detecting Workstation media, since their
volume IDs are truncated. Bug tracking that:

https://bugzilla.redhat.com/show_bug.cgi?id=1145264