rhel,script: Don't specify interface
Its not guaranteed that we'll have 'eth0' available or that it will be
named 'eth0' by kernel. Since kickstart allows us not to specify the
name of the interface, lets omit that and let kickstart choose the best
device for us.

This change has only been tested against RHEL 7.0. Would probably be
best to test against 6.x too.