fedora: Separate media entry for each ISO
Different ISOs have different volume sizes so to be able to add volume
size info to all medias, we have to create separate media entry for each
ISO. Moreover, apart from the fact that this makes one particular XML
file a lot bigger than we'd like it to be, this change makes a lot of
sense anyway.

This will also make addition of sequence numbers to multi-disc
installer medias, much easier later.