data: change PCI and USB device ID URIs
Currently PCI and USB device URIs look like

  http://pciids.sourceforge.net/v2.2/pci.ids/8086/2668
  http://www.linux-usb.org/usb.ids/80ee/0021

We want to associate filenames with URIs in a well-defined format.
These existing URIs, particular the PCI one, are too long in the
path component.

It is also desirable to use the industry standards body as the
domain name, rather than the domain of the specific project
we happen to import data from right now.

Thus the URIs are changed to

  http://pcisig.com/pci/1af4/1000
  http://usb.org/usb/80ee/0021

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>