debian: Update URLs to latest minor releases
During the lifetime of a major release, there are regular minor releases
of updated isos. Availability of ISOs for older minor releases seem not
to be guaranteed, leading to failures in test-mediauris (with
LIBOSINFO_NETWORK_TESTS set in the environment). This commit updates
debian 6/7/8 URLs to the latest minor update.