win8,installer: Tune based on variants
As Windows 8.1, Windows 8 (most likely) has to have the ProductKey
skipped during the WindowsPE phase of the script for the enterprise,
enterprise-debug and professional variants.

"Most likely" is said because I haven't been able to test the Windows 8
Professional variant as it needs a VLK and MSDN tells me that only
administrators can access keys in the Volume Licensing Service Center
(VLSC).

Also, I haven't been able to start an installation using none of the
debug ISOs, as the system reboots just a minute (or so) after booting
up. It seems to be a problem with the ISOs themselves, as I've
downloaded a few ISOs (different locales) and faced the same issue.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>