Fix COPYING file to refer to GPLv2+, not LGPLv2+
The LGPLv2+ is only applicable for code that needs to be linked
with other non-GPL compatible code. The actual osinfo database
files aren't linked in any traditional sense - they're merely
parsed. As such they can be just tagged with the GPLv2+, and
it would still be valid for them to be used for non-GPL compat
application code.

This also adds an explicit header to all the XML files to make
their license clear.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>