install-script,opensuse: Fix passwordless installation
While the passwordless installation worked properly with openSUSE Leap
42.1, it doesn't with current Tumbleweed and openSUSE Leap 42.2.

So, in order to allow the user log into the system using GDM, GDM's PAM
configuration had to be tweaked a little bit to allow passordless login
for a specific group, which has been added and contains the user as its
member.

As this solution also works properly for openSUSE Leap 42.1 I've decided
to not make it conditionally only to Leap 42.2 and Tumbleweed.

It's really important to mention that the solution has been described in
the following Arch Wiki:
https://wiki.archlinux.org/index.php/GDM#Passwordless_login

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>
Acked-by: Daniel P. Berrange <berrange@redhat.com>