rhel: set a volume size for atomic 7.0 ISO image
All the atomic ISO images have the same systemd-id and volume-id, so the
only way to distinguish them is based on volume size. The 7.1 and 7.2
media have a volume-size recorded, but the 7.0 media does not.

Currently we load entities in the order in which the filesystem returns
them from readdir(). If we're unlucky this can mean when matching media,
we'll match a 7.1 or 7.2 ISO against the 7.0 media by mistake.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Reviewed-by: Fabiano Fidêncio <fabiano@fidencio.org>