Add Red Hat Linux end of life dates
The historical RHL support policy was as follows:

 * Red Hat Support provides support for the two (2) most recent major
   versions of all software products.

 * All minor revisions of the current major version are supported. The final
   minor revision of the previous major version is supported.

Based on these rules, and the release-dates of each distro, we can
calculate the end of life dates for RHL 2.0 -> 6.1. For 6.2 onwards
the data comes from mailing list announcements. For RHL < 2.0, there
was no support, so the EOL dates are set to the release date of the
immediately following release.

Acked-by: Christophe Fergeau <cfergeau@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>