opensuse-unknown: Adjust the volume-id regex
The regex to match unknown openSUSE prereleases is becoming
bigger and harder to read.

Basically, what it's doing is:

(15\.[1-9]|(1[6-9]|[23]|[5-9])[0-9]\.[0-9]|4[3-9]\.[0-9])
---------   -----  ----------------   ---- -------------
    |         |         |                |       |
    v         v         v                |       v
  15.1+      16     20, 30, 50 - 90      v     43.0+
                                  .0, .1, ..., .9

I'm dropping both the 13.x series or older and 42.x series as those are
not going to have new releases anymore.

Also, (DVD|NET|Kubic) was introduced as this is the naming chosen for
DVD or NET installation or whether it's a Kubic  DVD installation.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>
Reviewed-by: Christophe Fergeau <cfergeau@redhat.com>