centos7: Add URL for CentOS 7
CentOS doesn't provide a direct link for downloading the ISOs. Instead,
their link redirects you to a mirror list, which is far from optimal for
our case.

After opening https://bugs.centos.org/view.php?id=15239, arrfab
suggested to use the symlink to the last respin which symlinks to
$latest.

This approach is the best we can take for now.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>