os: Add more ich6 annotations
Intel HDA was introduced in 2004, all modern OS support it out of the
box and it should be the preferred audio config with qemu. Add
more annotations for it.

win2k8 supports hda out of the box. Non-server windows already have
ich6 annotations.

For linux I used this rule: if the distro family already has an audio
device (ac97) specified somewhere in the hierarchy, stick the ich6
annotation beside either ac97 or virtio-block, whichever is specified
later. virtio-block first appeared in 2008 so I believe it's safe to
assume hda/ich6 support if virtio-block is present

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>