centos7: Adjust the "clones"
CentOS has taken the approach of not doing 7.x releases in the same way
RHEL does, but stick to "7" and just update new ISOs whenever they
rebuild the latest RHEL release.

It means that we will always have to update the "clones" in order to
point to latest RHEL release, otherwise the list of the devices won't be
up-to-date.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>