schema: Add support for pre-installed images
Let's create a new element called "image", which will be added to the
"os" element.

By adding the new "image" element, we'll start supporting pre-installed
images, at least in order to provide their link to Apps consuming
osinfo-db.

The xml entry will be similar to:
<os>
  <image arch="x86_64" format="raw">
    <variant id="cloud"/>
    <url>https://download/foo/bar.img</url>
  </image>
</os>

Mind that the "variant" is an optional element.

https://gitlab.com/libosinfo/osinfo-db/issues/10

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Pino Toscano <ptoscano@redhat.com>