schema: Add installer-script attribute to medias
Currently we've installer-scripts set for OSes and a single OS may have
different medias which only part of them actually support installation
via install-scripts.

This is the case, for instance, with Fedora where Live medias do not
support installation scripts.

In order to give the apps a proper way to check whether a specific media
supports or not the install scripts, let's have this new boolean
attribute set for each of the medias.

The 'installer-script' attribute is intended to be "true" by default on
every OS that has support for install-scripts.

The XML will look like:
<os>
  <media arch="x86_64" installer-script="false">
    ...
  </media>

  <media arch="x86_64">
    ..
  </media>

  <installer>
    <script id='http://foo.bar/distro/foobar/desktop'/>
  </installer>
</os>

In the scenario above the first media shouldn't be used for
installation using scripts while the second should.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>