schema: Add installer to the media
Having the installer tied to the OS is something that has been shown
not optimal, unfortunately.

An OS may have a few different medias and each media may support a
specific type of script or, even, not support an installer script at
all. With this in mind, let's have the installer scripts tied to the
medias themselves.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Christophe Fergeau <cfergeau@redhat.com>