ubuntu14.04: Copy installer under medias
Let's copy the installer under the medias as we've provided new APIs on
libosinfo for relying on the media content instead of relying on the OS.

The reasoning behind of that is because not all medias support
installers for both profiles.

Let's also leave the installer-script under the OSes as
osinfo-install-script relies on that.

The reason why the "desktop" profile has been left only in the OS is
because we can safely use the "desktop" script with the "server" medias
and have the very same result as when using the "desktop" medias.
Unfortunately, the opposite is not true (performing a "JeOS" installation
in a "desktop" media and having the same result).

Last but not least, we're just copying the installer under the medias
which the installer has been tested against.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>