ubuntu16.10: Copy installer under medias
Let's copy the installer under the medias as we've provided new APIs on
libosinfo for relying on the media content instead of relying on the OS.

The reasoning behind of that is because not all medias support
installers for both profiles.

Let's also leave the installer-script under the OSes as
osinfo-install-script relies on that.

Last but not least, we're just copying the installer under the medias
which the installer has been tested against.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>