win2k19: add Windows Server 2019 data
https://docs.microsoft.com/en-us/windows-server/get-started-19/whats-new-19

This new OS does not include any <media>, and this is done on purpose:
sadly, the ISOs have the same metadata (like Volume ID, etc) as win2k16
(Windows Server 2016), and thus it is not possible to distinguish them.
Theoretically, using the volume size can help with the detection;
however, the Windows ISOs have different supported languages, and there
is no way to specify the volume size for each language.

Regardless, having this OS is still useful for users to match an OS with
an osinfo ID.

Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>