mingw,spec: Remove obsolete Group tag
References: https://fedoraproject.org/wiki/Changes/Remove_Group_Tag

And "The Group: tag is not required in any live Fedora or EPEL release.
RHEL5 did need it, but EPEL5 did not as it was supplied automatically
via magic in the epel-rpm-macros package. The Packaging Guidelines have
indicated that the Group: tag should not be used since March of 2017."

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Pino Toscano <ptoscano@redhat.com>