test_isoinfo: Allow running a quickier version of the test
test_isoinfo takes a lot of time if running in the way it's been
implemented.

Let's add an environment variable to decide whether we want to run a
"quickier" or an expensive version of the test.

The quicker version would just test against the expected OS and ensure a
match happens, while the expensive version tests against all other OSes
to ensure we don't the media info does not match against any other OS by
mistake.

The environment variable added is OSINFO_DB_ENABLE_EXPENSIVE and it
expects an integer, as in 0 (or not passing it at all) to disable the
expensive test or non zero to enable it.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>