tests: util: Add DataFiles class
Use this class to hold all the file lookup helper routines, everything
currently in util. Init one global instance and use that in the test
cases. This will make it easier to cache lookup results as well

Make sure this is initialized from conftest, so if we hit any errors
it will bail out before running any tests. Use this to validate we
have a valid DATA_DIR too

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>