tests: Add custom pytest-3 --network-tests option
This wires up a --network-tests option to save us from needing the
OSINFO_DB_NETWORK_TESTS environment variable, though that method still
works. It works a bit differently in that unless specified it entirely
takes the tests out of the pool and doesn't list any SKIP options.

IMO this is nicer because it makes the default test suite output less
noisy. The network tests need to only be run periodically

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>