win7: Update pre/post install drivers
Let's update the pre/post install drivers to the latest ones.

Together with that, let's get rid of references pointing to Zeeshan's
personal page and let's use the official virtio-win fedora space (the
same one that will be used for Windows 8, 8.1, and 10).

By updating the drivers to the latest ones we get, for free, support to:
- virtio-scsi;
- virtio-rng;
- virtio1.0-*;
  *: net, block, console, rng, balloon, scsi, input

Note that the certificates have also been changed to the latest ones, as
those are needed in order to install those drivers.

Those certs can be obtained by:
- manually install a guest;
- install at least one virtio driver in the test VM (or install
  spice-guest-tools);
- run `certmgr.msc`
- goto "Trusted Publishers"
- goto "Certificates"
- right click in the Red Hat, Inc. certificate
- goto "All Tasks" -> "Export"
- save the file

https://gitlab.com/libosinfo/osinfo-db/issues/19

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Felipe Borges <feborges@redhat.com>