install-script,windows,desktop: Prevent infinite loop during post-install
If all the installation files generated are put in the same place, we'd
have both spice-guest-tools.cmd being called as expected but we'd also
call windows.cmd from windows.cmd, generating an infinite loop during
the post-install phase of the windows installation.

In order to avoid doing so, let's add a condition to *not* call
windows.cmd from windows.cmd file.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>