install-script,windows,desktop: Do not set user avatar for win8+
For Windows 8+ the image format and expected sizes of the image are
different than what's used for Windows 7 (or older).

This could be implemented (and then, re-enabled) in the future. But, for
now, let's just not set user avatar for win8+ as it's not going to work
there with the current state of libosinfo & consumers.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>