centos7: Rename CentOS 7.0 to CentOS 7
CentOS, differently than RHEL, does not support pointing releases at
all. If we knew that when adding the entry in the first place, we'd
never have a centos7.0 entry.

The easiest way to work around the mistake made in the past is by adding
centos7 as short-id so OSes with a "new enough" libosinfo could just
start using it without any issues.

Together with this change, let's also change the name from CentOS 7.0 to
CentOS 7 as it's the distro preferred way to advertise the system.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>