tests: Add test_os_short_ids
Let's test that all our short ids entry follow the [a-z0-9.-] pattern.

Mind that it's not enforced in our schema and I, sincerely, do *not*
think the schema should be changed as it may break user's own entries.
Even though, ensuring the pattern on upstream side is a good idea.

https://gitlab.com/libosinfo/osinfo-db/issues/31

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>