tests,test_ids: Take into consideration extensions
Extensions have their directories ending with ".d" and any name is a
valid name for the files inside this directory.

For instance, a fedora-30 extension would be:
/path/to/db/os/fedoraproject.org/fedora-30.d/memory.xml

This file is totally valid as long as memory.xml has its id as:
http://fedoraproject.org/fedora/30

This will be needed, later on, in order to validate more than OSes ids
as we rely on extensions for, for instance, our devices.

https://gitlab.com/libosinfo/osinfo-db/issues/33

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>