tests,util: Rename _ids_cb() to _shortid_ids_cb()
Let's add a _generic_ids_cb() method, to which is passed only the object
and the key and rename _ids_cb() to _shortid_ids_cb().

This change has been done as, in the future, we may want to parametrize
different entities type and those may use different keys (as filename
instead of shortid).

https://gitlab.com/libosinfo/osinfo-db/issues/33

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Cole Robinson <crobinso@redhat.com>