endlessos: autogenerate eos-3.X.xml.in
Endless OS publishes a new release every month. Every six months we make
a major release (3.X.0); the intervening releases are point releases.

Endless operates a rolling release model: every release since 3.0.0 will
automatically upgrade to the latest 3.X.Y release once it connects to
the internet. We don't expect this to change; if it does, we'll switch
to an Endless OS 4 branch.

We define an <os> per 3.X release series. As new point releases on that
branch are published, the script will update the image URLs to refer to
the latest release. When a new release series appears, the previous
series is marked as EOL as of the date of the first release in the new
series.

The data is generated based on our published index of OS images. This is
the same index we use to generate the links on our website, so it is
very likely to be accurate.

Ideally at some point in future, this script could be run and its output
pushed into a new osinfo-db release without human intervention; in the
meantime, we'll run it and send a MR when we make a new release.

Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>