centos7: Add specific entries for minimal variant
CentOS 7 is almost a rolling project, in a way that the project prefers
to *not* expose its minimal versions to the users as any update would
take the users to the latest released version.

Unfortunately, its medias have the very same volume-id making it quite
hard to differentiate the medias unless we use their volume-size.

This patch adds the specific media based volume-sizes entries for the
minimal variant.

Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>