install-script,opensuse: Drop bootloader section
AutoYaST already takes care of selecting the most appropriate
configurations without us having to (try to) tweak it.

As a side effect of this commit, the loader type will be selected by the
installer accodingly, meaning that we don't have to deal with BIOS / EFI
boot loader on our side.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Imobach González Sosa <igonzalezsosa@suse.com>