po: add custom ITS rules
These ITS rules extract name and vendor tags from the libosinfo XML
files. At the moment, the .in files are targetted, so the _ variants of
the tags are extracted.

The path of the file is chosen so XDG_DATA_DIRS will point to
$VPATH/po: this is needed because the join mode of itstool honors the
ITS rules passed as -i only since version 2.0.6.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>