silverblue,install-script: Rework user creation
Silverblue, till its "30" release, allowed us to:
- set passwordless accounts;
- set root password;

Those things were just mimic'ing Fedora, and done during the
"post" phase.

However, since its "31" release, it's been broken as the user was
created but no password was set;

In order to solve this, the easiest solution is stop doing such
things during the "post" phase, and use the "kickstart" way. The
drawback, though, is that we have to require user / admin password
for such case and this is not something we can conditionally do
depending on the distro's version.

Summing up, as the passwords will be required, let's switch to
only using kickstart methods for setting up the user and admin
accounts / passwords always.

Mind tho, that for Silverblue 31 and onwards, the root password
won't even be used as `su` doesn't seem to be the way Silverblue
wants their users becoming privileged (`sudo` would be the
 preferred way).

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>