os: Add generic linux2020
This is a generic entry representing a rough linux defaults for a
distro released in 2020.

Its device and resource values are derived from Ubuntu 20.04 LTS.

Use an ID based on libosinfo.org to represent that this is not a
real distro but a synthetic db entry.

Some of the rationale for this was discussed here:
https://www.redhat.com/archives/virt-tools-list/2020-September/msg00017.html

Signed-off-by: Cole Robinson <crobinso@redhat.com>