debian10: Add "generic" cloud images
Although there's no direct link for the "latest" or "current" release,
the links as they are do not change nor get removed when a new image is
released.

With this in mind, it seems okay to add the info about those here and do
an update before releasing osinfo-db.

The definition of the debian images can be found
https://salsa.debian.org/cloud-team/debian-cloud-images and does
mention:
* generic: Should run in any environment
* genericcloud: Should run in any virtualised environment. Is smaller
                than generic by excluding drivers for physical hardware.
* nocloud: Mostly useful for testing the build process itself. Doesn't
           have cloud-init installed, but instead allows root login
           without a password.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>