tests: retry 3 times before failing the URL test
Currently we're facing a few issues with the URL tests, which has been
false-positively failing for a long time.  This happens as some distros
URLs are flaky, and retrying the test a second time is usually enough to
have it passing.

Knowing this, let's retry 3 times before actually failing the test,
decreasing then the amount of false-positives we have.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>