install-script,autoyast: Don't try to set the user's avatar
This is causing issues when using an OS with a different language than
English (es_ES, in my case).

The error reported is:
```
Language 'es_ES.UTF-8' was not found within the list of supported
languages avaiable at /usr/share/YaST2/data/languages.

Fallback language en_US will be used.
```

This led me to think that the problem could be in the language we set
for the system, but after several tests I found out that no, it was not.

Removing the bits related to the user avatar solves the issue.

Fixes: #77

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>