install-script,autoyast: Set "openSUSE" as a product
openSUSE Tumbleweed ISOs now allow the user to either install
Tumbleweed or MicroOS, without fallbacking to one of those by default.

While we still don't support MicroOS, the safest option for us is to
always set the product as "openSUSE".  Whenever MicroOS is added to
osinfo-db, we use the distro name to differentiate the product to be
used.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>