install-script,autoyast,jeos: Don't explicitly install dhcp-*
dhcp-tools is not needed, and actually no longer present on openSUSE
Tumbleweed.

dhcp-client will be pulled in automatically as a dependency of
Networkmanager.

Removing those allow us to not complicate the scripts without a good
reason to do so, as we'd already have make a special case for
tumbleweed, due to the dhcp-tools removal

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>