data: add minimal record for CentOS Stream 9
This provides the bare minimum record for CentOS Stream 9, with data
copied from CentOS Stream 8. All media and all treeinfo records except
the generic one are removed, since nothing is officially published
yet. This record is good enough for virt-install to be able to do
the right thing when launching the CentOS Stream 9 installer.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>