tests: check for content-type of initrd URLs
Make sure that the content-type received for URLs of ISOs is a possible
acceptable one; the currently accepted types are:
- application/octet-stream: an unspecified/generic sequence of data
- application/x-gzip: gzip-compressed initrd files

Not all the servers return a content-type for initrd files, so this
applies only to URLs with a returned content-type.

Though, even with the above exceptions, we should still catch very big
mistakes like HTML pages for errors or redirections (text/html).

Signed-off-by: Pino Toscano <ptoscano@redhat.com>