Add Windows 11 to the database
Some things to note

 - It mandates UEFI for booting
 - It mandates a TPM
 - It mandates very new Intel / AMD CPU models
 - The ISO PVD metadata for the image is indistinguishable
   from the Windows 10 entry we have.
 - Min RAM is 4 GB, disk is 64 GB.

This adds an entry for UEFI firmware, and reports that BIOS is not
supported, as we don't expect users to do regedit hacks during the
install process.

We haven't tracked support for TPMs in any way historically, so nothing
is done in this respect, but we're going to need to figure out an answer
there to help management apps since none provision a TPM by default.

We don't have any way to represent CPU generation requirements and
this is probably an impractical thing to solve since most OS do not
give any fine grained details. Microsoft happens to give an enourmous
list of CPU model identifiers

  https://docs.microsoft.com/en-gb/windows-hardware/design/minimum/supported/windows-11-supported-intel-processors

but I don't want to copy that into libosinfo.

There's no practical way to avoid the clashing ISO PVD matches
between Windows 10 and 11. With current libosinfo APIs it will
be non-deterministic which OS matches, so users will have to
apply an override if relevant. Libosinfo will need new APIs to
let apps get a list of multiple matching OS, whereupon they can
prompt users to select the specific one, or simply decide to use
the most recent.

Since this introduces ambiguity for Windows 10 vs 11 ISO images,
symlinks are added for the test suite data files between the
different versions. The real file is the intended match, and
the symlinks are the accidental incorrect matches.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>