tests: check for content-type of image URLs
Make sure that the content-type received for URLs of images is a
possible acceptable one; the currently accepted types are only for qcow2
images:
- application/octet-stream: an unspecified/generic sequence of data

Not all the servers return a content-type for images, so this applies
only to URLs with a returned content-type; in particular, there is
currently no returned content-type for raw & vmdk disk images.

Though, even with the above exceptions, we should still catch very big
mistakes like HTML pages for errors or redirections (text/html).

Signed-off-by: Pino Toscano <ptoscano@redhat.com>