os: Add generic linux2022
This is a generic entry representing rough linux defaults for a
distro released in 2022, similar to other entries like linux2020.

Its device and resource values are derived from Ubuntu 22.04 LTS,
with the exception that virtio1.0-gpu is re-listed as supported.
It is disabled for Ubuntu 22.04 due to distro bugs, but other distros
from this timeframe are expected to work well with virtio gpu.

Signed-off-by: Cole Robinson <crobinso@redhat.com>