data: switch to images/pxeboot for Fedora unknown
Copy of c9eb52afff644c2768674e044498c5908dab5629, but for Fedora
'unknown' instead of 'rawhide'.

Fedora 37 rawhide has dropped syslinux in favour of grub for ISO
boot and as such, the 'isolinux/' directory no longer exists.
The kernel/initrd files in that directory held 100% identical
content to the files in 'images/pxeboot/', so this is a straight
swap.

https://bugzilla.redhat.com/show_bug.cgi?id=2103835
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>