tests: check .treeinfo files
In case a <tree> has <treeinfo>, it means the details of that tree are
in a .treeinfo file. Hence, if there is an <url> set for that tree,
check that there is a .treeinfo file on the remote location.

Also check the content-type for them; right now, only the generic
application/octet-stream is returned for them by some servers.
Still, it is enough to catch very big mistakes like HTML pages for
errors or redirections (text/html).

Signed-off-by: Pino Toscano <ptoscano@redhat.com>