tests: read .treeinfo from open file object
We open the file object for a treeinfo file, so load it from that using
read_file() rather than making ConfigParser open the file again with
read().

There is no behaviour change, only a very minor optimization, and the
'out' variable is actually used now (fixing the flake8 issue F841).

Signed-off-by: Pino Toscano <ptoscano@redhat.com>