endlessos: Don't create empty isodata files on error
Previously, if `isoinfo` was not installed, the target file would be
created, empty. Re-running the script would not fix this problem because
fetch_isodata() early-returns if the file exists.

Instead, write the output of `isoinfo` to a temporary file, and link it
to the destination file on success.

I tried using an unnamed temporary file, then linking
"/proc/self/fd/$FILENO" to the target file, and couldn't get it to work.
This is good enough.

Signed-off-by: Will Thompson <wjt@endlessos.org>