tests: inline parameter tuples when possible
Instead of having a tuple as parameter to certain parametrized
functions, then declare the arguments to pytest.mark.parametrize(),
and pass the values directly to pytest.param().

In case of test_urls, the first argument (the OS short-id) was used only
as ID for the parameters, and thus it was dropped.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>