tests: collect & test URLs by schema
Collect the URLs for each OS according to their schema: http/https,
docker, and others. This will makes it easier to do proper per-schema
processing/testing, and in case add tests for other schemas in the
future.

Since _check_url() is used also for testing the underlying http URL for
a docker URL, add a parameter to log the URL tested; the logging of the
HTTP request will show anyway the real URL.

Even though there are no supported URL types other than HTTP and docker
ones, log them as warning, and make the test fail: this is what they
would do anyway currently, so there is no behaviour change; this also
makes it easier for us to notice the situation, and act on it.
There are no such URLs in osinfo-db as the time of this commit.

Also, this decouples the schema of an URL from its type; there was an
implicit assumption that containerdisk URLs were docker URLs.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>