tests: refactor UrlType
Instead of a simple enum of int values (which have no usefulness),
change UrlType to have the content types of each type as value: this
makes each type and the content types for it together, harder to
mismatch.

The only gotcha here is working around the duplicate values handling of
enum.Enum: by default duplicate values are turned into aliases, whereas
we want each type to be unique, no matter its content types. To overcome
that, create UniqueSet as small subclass of set that never compares
equal to any other, including another set with the very same values.

As negative side effect of the "never compare equal" values is that enum
equality comparison is not possible directly, so the only way is to
compare their names instead.

As related side effect, now _is_content_type_allowed() is much smaller,
and does not need to be updated anymore for new types.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>