tests: check for content-type of trees
Make sure that the content-type received for URLs of trees is a possible
acceptable one; an HTTP request for a tree usually returns an HTML page
with the file listing, so text/html is the content-type that needs to be
allowed for this type of URLs. Still, even with the risk of receiving an
HTML error page, it is some small improvement.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>