tests: check for content-type of vhdx disk images
Make sure that the content-type received for URLs of vhdx disk images
is a possible acceptable one; there are no vhdx disk images in osinfo-db
at the time of this commit, so no content-types are set. In case any
will be added in the future, it will be easy to add content-types for
them, if returned by the servers.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>