tests: check ISO languages using .lng files
Among the isodata test files, there are also ".lng" files associated
with the ".txt" ISO files: they are used to specify the languages of an
ISO file; this feature is mostly used for Windows ISOs, and it was
introduced in the libosinfo test suite long ago [1] (then copied over to
osinfo-db).

Make again use of these data in test_isoinfo:
- for each isodata file, read the languages available for it, if present
- in case there are languages and the ISO matches the wanted OS, try to
  read the languages from the <iso> definition for the actual ISO using
  the same logic used in libosinfo, and then compare it with what is
  expected

All the existing ".lng" files work fine, no regression was found.

Fixes: #38

[1] https://gitlab.com/libosinfo/libosinfo/-/commit/ac8c5ef7f57505dbd68d9094c0c3e6de247478b7

Signed-off-by: Pino Toscano <ptoscano@redhat.com>