tests: add a basic generation test for install scripts
Add an initial test to check that the generation of install scripts
actually works; this is done by doing the following steps:
- for each OS, iterate through the medias and the trees that have an
  install script specified
- for each media and tree, generate the config XMLs for all the
  injection methods available
- try to transform each of the config XMLs using the template of the
  install script
- do a basic validation of the result:
  - for XML outputs, check it can be parsed as XML
  - for command line outputs, check it is not empty and has at most one
    new line at the end

The parameters are filled from dummy/sample values for the required
ones, or for the ones provided by default by libosinfo.

The current way tests all the possible combinations of
  media/tree x install script x injection method
which may be a bit wasteful, however it is better than potentially
skipping needed combinations.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>