tests: try to validate AutoYaST XMLs
Try to validate AutoYaST XMLs, i.e. the configuration files for
automated installations using AutoYaST (typically on SUSE systems):
this is done by using the RNG schemas available for AutoYaST, either as
available from the yast2-schema-collection SUSE package, or by a local
checkout of it.

Add the AutoYaSTSchemas class to lookup the schema depending on the OS
and the architecture, caching the results; because of this, expose a
shared instance of it as fixture to the "test_generate_scripts" test.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>