Revert "tests: Make test results more readable by reducing tracebacks"
The problem is that we *do want* tracebacks. When running tests,
especially in CI contexts in which all you have are the results, not
having the full traceback printed/available means that it is harder to
find out what triggered a certain error. The errors ought to be
debuggable by default, and you want to run them locally in a different
way to have less details, you can always run pytest manually.

This reverts commit 8958ef479bbe9eacd87ccc09001624a572a8a733.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>