scripts: Unbreak osinfo-db-add-iso.py
After "d384ddb0 tests/test_isoinfo: convert to use 'pathlib'
module" the get_isodatamedia() accepts a Path instead of plain
str. Fix osinfo-db-add-iso.py so that it passes expected type.
Also, in the same commit the .filename attribute of _ISODataMedia
class was renamed to .path. Reflect that in the script too.

Fixes: d384ddb0b115f4afd87ee27d5a22bf47ced7e508
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>