Rocky Linux: Change URLs of ISOs to concise name
Release Engineering will be deprecating the use of Rocky-ARCH-TYPE.iso
in a future release, as it was determined to cause confusion not only by
members within the project, but also be the end users. It was also
determined that it was also a duplication of the same effort as
Rocky-X-latest-ARCH-TYPE.iso, which has existed for some time.

Historically Rocky-X-latest-ARCH-TYPE.iso has existed to fulfill the
need for having a pointer to the latest ISO in the event they had to be
rebuilt but also when a new release arrives. It points to the absolute
latest at all times.

This helps provide the same idea for both the end user and also
osinfo/libvirt.

Signed-off-by: Louis Abel <tucklesepk@gmail.com>