gentoo: Raise minimal requirements and add recommended ones
There an issue filled against libvirt that basically boils down
to not enough memory being set by default for Gentoo live ISO.
No surprise there - we claim the minimal requirement is 256MiB
(!). And live ISO starts a KDE session on boot.

Now, it's hard to come up with something that'll work in 100%
cases and still be minimal. There's this recommendation in
Gentoo's handbook [1], but the page was last edited in 2014.

Anyway, I've tried to boot the live ISO and the VM was using
~1.5GiB of memory. So let's put the minimum as 2GiB as it could
grant enough memory for emerge/compilation or basic system tools.

And while at it, recommend 4GiB of RAM and 20GiB of storage.
There are packages that require much more, e.g.
www-client/chrominum may require up to 16GiB [2], but that's
rather exception. And recently, Gentoo added support for binary
packages too which helps to save some memory as users don't need
to compile packages themselves.

1: https://wiki.gentoo.org/wiki/Handbook:X86/Full/Installation#Hardware_requirements
2: https://github.com/gentoo/gentoo/blob/master/www-client/chromium/chromium-125.0.6422.14.ebuild#L352
Resolves: https://gitlab.com/libvirt/libvirt/-/issues/600
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>